Changing SearchStats.playTime from int to unsigned long.
git-svn-id: https://svn.musicpd.org/mpd/trunk@5951 09075e82-0dd4-0310-85a5-a0d7c8717e4f

+++ src/dbUtils.c
@@ -41,7 +41,7 @@ typedef struct _LocateTagItemArray {
 typedef struct _SearchStats {
 	LocateTagItemArray locateArray;
 	int numberOfSongs;
-	int playTime;
+	unsigned long playTime;
 } SearchStats;
 
 static int countSongsInDirectory(int fd, Directory * directory, void *data)
